storage_controller: Add `public_key_path` argument to read key from file
Add a command line optional arugment, `public_key_path`, which is the
path to the public key file for JWT authentication. This helps simplify
the deployment of the storage controller to match the other services,
where the public key is read from a file rather than inline on the
command line. In addition, this allows the JwkAuth::from_key_path to
load multiple public keys when this path points to a directory, or if it
points to a jwk.json file containing multiple JWKs.